feat: split swc test
